requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
openai==1.40.0
streamlit==1.28.1
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
beautifulsoup4==4.12.2
aiohttp==3.8.6
asyncio==3.4.3
openpyxl==3.1.2
markdownify==0.11.6
jinja2==3.1.2
